carrito

+++ application/controllers/Compras.php
@@ -24,18 +24,28 @@ class Compras extends CI_Controller {
 
 
     function compra($id)
-    {
-        echo 'se ha comprado el producto con id '.$id;        
+    {   
         $articulo = $this->Tienda_model->datos_libro($id);
-        echo "<pre>"; print_r($articulo); echo "</pre>";
-        $articulo['id']=$id;
-      
-        $this->Carrito->add($articulo);    
-        $res = $this->Carrito->get_content();
-        
-        //echo "<pre>"; print_r($res); echo "</pre>";
-        $carro = $this->load->view('Carro', Array('articulos' =>$res),true);  
-        $this->load->view('Plantilla_carro',Array('carro' => $carro));
+        if($articulo['cantidad']>0) //si hay existencias operamos
+        {
+            //restamos en 1 el stock del articulo
+           
+            //$this->Tienda_model->setLibro($articulo['id'],$datos);
+            echo "<pre>"; print_r($articulo); echo "</pre>";
+            $articulo['id']=$id;
+            $articulo['cantidad']=1;
+
+            $this->Carrito->add($articulo);    
+            $res = $this->Carrito->get_content();
+
+            //echo "<pre>"; print_r($res); echo "</pre>";
+            $carro = $this->load->view('Carro', Array('articulos' =>$res),true);  
+            $this->load->view('Plantilla_carro',Array('carro' => $carro));     
+        }
+        else {
+             $carro = '<div class="alert alert-info">Lo sentimos no quedan existencias de ese articulo :(!</div>'; 
+             $this->load->view('Plantilla_carro',Array('carro' => $carro));  
+        }
     }
     
     function vaciar()
@@ -48,6 +58,9 @@ class Compras extends CI_Controller {
     function eliminar($pro)
     {
         $this->Carrito->remove_producto($pro);
+        $res = $this->Carrito->get_content(); 
+        $carro = $this->load->view('Carro', Array('articulos' =>$res),true);  
+        $this->load->view('Plantilla_carro',Array('carro' => $carro));
     }
     
 }
